api/requirements.txt
fastapi
tortoise-orm
typing_extensions
bcrypt
iso8601
PyJWT
aiomysql
pytz
uvicorn
passlib
python-multipart
